requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
Flask-WTF==1.2.1
frozenlist==1.4.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.28.1
Pillow==10.1.0
requests==2.31.0
tqdm==4.66.1
urllib3==2.0.7
uvicorn==0.24.0
Werkzeug==3.0.1
WTForms==3.1.1
yarl==1.9.2
